feat: Remove Tomcat version in default 404 page
Present in the template server.xml meant for configuring SSL. If a user
were to mount this file without the change over the server.xml that is 
present in the Tomcat image, then it would undo the changes and the 
Tomcat version would remain visible on the 404. We want this behavior 
to be there by default.