feat: switch to environment to YAML dictionary
The dictionary format in YAML allows for less possibility of future users being confused on the parsing of their value for an environment variable. Because the list format currently treats the key,value pair as one whole string, where as the dictionary keeps k,v separate. You can quote your value in a dictionary purely for semantics and not have it be included in the actual value (unlike list).