feat: Added scripts for backup and restore and docs
Rough draft of what it should look, verified it works for the expected use
case when there's no customizations (e.g. remote db, different appdata
path, etc.)

Central idea is copying the volumes that the tomcat and db containers
use and storing it as a backup. Then when the restore script is run,
that backup name is provided and the old volumes are replaced with the
backup volume.

Currently, the backup script depends on the containers running. Will
probably change this because I'm unfamiliar with the possibility of
there being intermediate data commited to the db while this copy of the
volume is built. Will use the same approach as the restore script where
a new volume is built when the containers aren't running probably.

Docs will need changing, as I wrote them before I actually started
implementing the scripts, so some expected inputs from the doc version
to the actual script differ.

Bash versions need to be made, which will mostly still be the same as
the powershell scripts except for how the parameters are passed in and
we'd use echo instead of Write-Verbose.