Improve backup & restore comment based help
get-help -full will now return helpful information on the parameters of
these scripts and how to populate them if needed. Also links back to the
docs.

Added check in the restore script if the volumes actually exist before
they're deleted so it doesn't error in a case where a user may have done
docker-compose down -v and is trying to run the restore script.