Refactored restore script, updated docs, improved output messages
Change prompt logic to instead be testing when the user said yes and then
else no. Added case insensitivity clarification in comment

Made console messages more consistent with ellipsis when work is about
to occur

When overwriting backup volumes, remove existing archive files (e.g.
user w external db overwrites an old backup that had both appdata and db
archives so it now only contains an appdata archive)

Removed UseExternalDb switch in the restore script since it can be
inferred from the backup volume by looking to see if a db archive
exists. Updated documentation so it no longer includes this switch for
the restore script

Changed line endings on scripts to LF to facilitate passing multiline commands
to the ubuntu container when working with volumes

Changed logic for restore script when specifying AppDataVolumeName so
the explicitly set value is always used. Previously, when using an
external db, if the default value was explicitly set for the
AppDataVolumeName parameter it would be overwritten with the default
AppDataVolumeName value for the external database configuration
(codedx-appdata-ex-db-volume). It's possible a user may want to only
replace the appdata, despite not using an external database. An appdata
only archive could be created with the backup script by specifying
-UsingExternalDatabase. Then they restore it with the command `restore -BackupVolumeName appdata
-AppDataVolumeName codedx-docker_codedx-appdata-volume`. With the previous approach
when it sees there's no db archive and checks if the appdata volume name is the default value it will
replace it with the external db default value. This was unintended
behavior. By using the built-in $PSBoundParameters map we can tell if the parameter
was explicitly set.